Rearchitecting `skpro` on `skbase` - part 1: CI-CD, base classes, package skeleton (#9)
This PR is a major rearchitecting of `skpro` based on `skbase` and the
`sktime` probabilistic prediction framework. Contents:

* CI-CD is using the `sktime`/`skbase` templates
* folder structure follows architecture of `sktime`, `skpro`
* base classes and testing framework used from `skbase`
* new probabilistic supervised regression base class design for the base
regressor
* `BaseDistribution` as return type, same as in `sktime`
* an example class which has been fully rearchitected, the residual
double regressor, `ResidualDouble`

Overall, this is hopefully much more streamlined and extensible.

Legacy code has been moved into the new folder structure, and is
expected to be migrated eventually to the new framework. Legacy files
have a `LEGACY MODULE` comment at the top.

Linting has been applied, but there are still some pydocstyle failures.
These will be addressed as part of the migration, and should not be
highlighted in future PR unless one of the affected legacy files is
changed.