[ENH] integrate `check_estimator` with `TestAllEstimators` and `TestAllRegressors` for python command line estimator testing (#138)
This PR properly integrates `TestAllEstimators` and `TestAllRegressors`
with the `check_estimator` utility and fixes. These were so far not
connected due to dependencies on the test scenarios which now have been
added (and a bug in `scitype` name resolution).

The integration happens in three places:

* `TestAllEstimators` now inherits from `skbase` `QuickTester` to allow
command line testing
* `check_estimator` now runs `TestAllEstimators` for estimators
* `TestAllRegressor` integration with `check_estimator` is fixed, the
scitype string was erroneously `"regressor"` instead of, correctly,
`"regressor_proba"`.