[ENH] support for survival/time-to-event prediction, statsmodels Cox PH model (#157)
This PR implements framework support for survival (aka time-to-event or
failure time) prediction, adds tests, and an interface to `statsmodels`
cox proportional hazards models as test case.

Depends on https://github.com/sktime/skpro/pull/155 and
https://github.com/sktime/skpro/pull/159 which should be merged first.

### Design

Survival prediction models use the current `BaseRegressorProba` base
class, which has `fit` extended to take a third argument `C`, a
dataframe-like with a censoring indicator.

Regressors capable of making use of the third argument `C` are
identified via the `capability:survival` tag (being `True`). Regressors
without this tag also take `C` but ignore it, corresponding to the
"ignore censoring" reduction strategy.

This way, all existing regressors can be used for survival prediction
and vice versa.
The interface is also fully downwards compatible for users - `C`
defaults to `None` - and for extenders - estimators without the tag do
not assume a `C` present in fit, as in this case only `X_inner`,
`y_inner` are passed in `fit`.

As the `predict` and `predict_proba` interfaces remain unchanged,
metrics do not need to be adapted, they directly work.

To avoid cluttering the docs for users who are interested primarily in
probabilistic regression without censoring, models with the
`capability:survival` tag have a more detailed `fit` docstring. The
difference is mediated via a base class `BaseSuvReg`, which is the same
as `BaseRegressorProba` with docstring overrides.

### Testing

As time-to-event models inherit from `BaseProbaRegressor`, the existing
`TestAllRegressors` suite tests runs on all survival prediction models.

A scenario with a non-trivial `C` is added.

As regressors and time-to-event models have an interchangeable interface
(see above), both are tested with non-trivial `C`, with `C=None`, and
without a `C` being passed.

### Further contents

* an inteface to `statsmodels` proportional hazards models,
`skpro.survival.coxph.CoxPH`, to showcase and test the interface
* `Pipeline` is updated to accommodate survival models, for this the tag
needs to be carried and `C` passed through in `_fit`
* update to the API reference - new page for survival prediction
* survival prediction extension template