[ENH] distributions: `pmf` and `log_pmf` method (#295)
This PR adds a `pmf` and `log_pmf` method to the base interface. Fixes
#289

In accordance with #229, these return 0 resp `-np.inf` if the
distribution is continuous.

Also makes the following, connected changes:
* `pdf` return 0 for discrete distributions
* removes the discrete/continuous handling logic from the `scipy`
adapter, as this is now in the base class

I've also changed the way in which `TestScipyAdapter` queries the
distributions - by inheritance, not by tag. This is since the tag is
"mechanical" (for internal testing only) and it might confuse users to
see a value in `object_type` which is not related to an external API
property.