[MNT] deprecation handling for `CyclicBoosting` (#329)
Deprecation messages for `CyclicBoosting` for deprecations and changes
in 2.4.0.
Alternative to #320.

* restores sequence of parameters to sequence in version 2.2.2, to avoid
breakage in the 2.3.0 release
* adds deprecation message for `bound` parameter
* renames new `version` parameter to `dist_type` - the name is a bit
misleading, as it will probably imply "version" in the semantic
versioning sense to an ordinary user. Does not require deprecation,
because added since 2.2.2.

Also makes improvements to the docstring of the `CyclicBoosting`
estimator.

Difference to #320: does not carry out deprecations for distributions,
under a working assumption that #327 will be merged.