[BUG] fix `test_methods_p` logic when `shuffle` is `True` (#381)
fixes #380 

#### What does this implement/fix? Explain your changes.

`object_instance` is the distribution before shuffling.
`d` is the one after shuffling so it needs to be used instead 
`res=getattr(object_instance,method)(p)` to be replaced with `res =
getattr(d, method)(p)` as this will account for the object_instance
after shifting the distribution. This will ensure the outputs are
checked with the shifted distribution's instance.