[ENH] Added Inverse Gamma distribution (#415)
#### Reference Issues/PRs
Related to the development of BayesianLinearRegression class. (#358)

#### What does this implement/fix? Explain your changes.

This pull request implements the Inverse Gamma distribution and Beta
Distribution as a new distribution class.

The Inverse Gamma implementation wraps around `scipy.stats.invgamma` to
provide functionalities consistent with other distributions in the
library. I introduced this distribution as the conjugate prior for the
variance (σ²) of a Normal distribution for subsequent use in the
`BayesianLinearRegression` class. (#358)