[ENH] Syncing datatypes module `_check.py` and `_convert.py` with `sktime` (#432)
As per requested by @fkiraly, we are updating the files `_check.py` and
`_convert.py` to match functionality with the `sktime` library. Note
that `_common.py` is exactly the same for both libraries, so no changes
were needed.

This pr also should allow list of strings to be accepted as valid
`X_inner_mtype` and `y_inner_mtype`.