[ENH] Skew-Normal Distribution (#512)
#### Reference Issues/PRs  

Fixes #510  

#### What does this implement/fix? Explain your changes.  

This implementation adds a new `Skew-Normal` distribution as a
probabilistic model in `skpro`. The inspiration comes from the paper
(https://www.ine.pt/revstat/pdf/rs130105.pdf).

The `Skew-Normal` distribution captures skewness in data, improving the
flexibility of predictive distributions. Scipy already implements this
as `scipy.stats.skewnorm`, which is used as a dependency to avoid
redundant code.