[BUG] fix `MapieRegressor` dependency check, and `BaseProbaRegressor._predict_proba` default (#540)
This PR fixes two unreported bugs:

* the `MapieRegressor` dependency check was always failing due to the
capitalization of `MAPIE` on pypi - not `mapie`. This led to the
estimator not being useable since the upgrade to strict PEP 440 (which
included capitalization sensitivity) of the `python_dependencies` tag.
* the `BaseProbaRegressor._predict_proba` sensible default would not
work properly if the expected return of `predict` was not a
`pd.DataFrame`. This has been remedied by using the idiomatic method
`_get_columns` to determine the column names.