[BUG] correct warning message for distribution `cdf` in approximate case (#556)
When a distribution `cdf` does not have an analytical implementation,
the fallback implementation from the `BaseDistribution` may be called,
which is a simple Monte Carlo estimate that will trigger a warning.

However, the warning refers to the mean, and not the cdf - this is a
mistake.

This PR fixes the error message.